Cli (#35)
* leafable hasher to be used for hashmap

* make public salt inner

* merkle proof visibility

* from_hash_out extra check

* del .DS_Store

* add test

* delete fraud prover

* rename is_valid -> did_return_sig

* add dummy

* spent witness

* error handling when initiate witness

* fix tx inclusion test

* fix prev_account_tree_root genesis value

* delete get_genesis_block_hash test

* wip

* clean balance

* impl trim account ids

* clear

* impl post_registration_block

* mock contract

* impl SyncValidityProver

* fix block builder

* wip

* onelock

* tree serialization

* serialization of user data

* save and get

* wip

* impl get_transition_data

* add getter & setter balance proof

* wip

* wip

* impl process deposit

* process tx

* impl process transfer

* balance logic

* wip

* add validation

* wip

* strategy

* generate strategy

* wip

* sync balance proof

* wip

* improve deposit

* wip

* wip

* common tx separation

* generate tx

* wip

* impl client

* e2e deposit

* wip

* wip

* wip

* wip

* print balances

* wip

* wip

* fix transfer padding

* wip

* delete wallet

* wip

* fix query bug

* wip

* withdrawal

* wip

* save withdrawal data

* chore

* make private iner validity prover

* delete send witness

* withdrawal separation

* signle withdrawal

* fix e2e

* change wrap timing

* rename

* minor fix comment

* sender leaves

* block builder 2step

* minor fix

* fix verify signature

* fix

* change order of backup

* wip: store vault server

* wip

* add withdawal sync

* wip

* wip

* wip

* update lpt

* make public indexed merkle leaf

* validity circuit -> validity vd

* validity circuit to verifier data

* block builders flow updated

* minor fix

* minor change interface

* from siblings

* wip

* wip

* delete recursivable

* add generate_from_provisional

* add serialize

* add serde

* serialize

* add clone

* add clone

* clone

* fix block builder

* remove tx root duplication check

* add reset to block builder

* append processed

* assert to ensure

* fix utils

* fix trees

* fix ethereum types

* fix common

* fix circuits

* fix

* fix overflow

* add reset

* add reset

* make public data of WithdrawalWrapperCircuit

* hide mock

* make public signature rand

* make public signature rand

* make index public

* hash to weight

* serialize & deserialize merkle tree

* make private packed tree

* add serialize to hashout

* fix not dummy

* Update zkp (#29)

* fix unconstrained membership proof (#16)

* fix unconstrained membership proof

* add comment

* fix unconstrained pubkey in update circuit (#17)

* update bn254 library (#18)

* connect block pis (#19)

* fix genesis value of public state (#20)

* safe_split_lo_and_hi (#21)

* safe_split_lo_and_hi

* update comment

* add rnage check to hash_out_target of KeccakLeafableHasher (#22)

* fix misleading naminig of prev_block_number (#23)

* fix misleading comment (#24)

* update plonky2_bn254 & hash_to_weight_circuit (#25)

* assert  prev_balance_pis.pubkey == pis.pubkey (#26)

* add next_account_id (#27)

* add next_account_id

* add assertion of incremental account_registration_proof.index

* upadte plonky2_bn254 dep

* fix get_genesis_block_hash

* fix 5.1

* fix 5.2

* Add posted time in block

* Fix block time

* fix flag

* exclusion for did not returned sig

* add prev_private_commitment but test fails

* delete account registration or dummy

* move block builder

* fix e2e test

* comment out wallet

* add u64

* revert block time

* add timestamp to blocks

* add timestamp to public inputs

* add expiry to signature

* add depositor and nonce to deposit

* wip

* add expiry to sign message

* construct validity witness

* test  utils

* witness generator

* validity state manager

* update proof

* fix balance_processor_receive_transfer

* fix balance_processor_receive_transfer

* balance_processor_deposit

* transfer_inclusion

* wip

* wip

---------


* from u64

* delete nonce

* fix rand sig

* fix update logic

* Last tx hash mismatch from assert eq to tx hash

* is_eligible

* recover withdrawal processor test

* add directory

* pretty

* Claim circuit (#30)

* determine lock time

* add mining nullifier

* StartTimeValue

* StartTimeTarget

* set witness

* start time circuit

* fmt

* clippy wip

* single claim proof

* single claim proof

* hash inner

* hash inner circuit

* path change

* fix mock

* rename

* deposit_time_proof

* single claim proof

* wip

* U64 to u64

* save claim circuit

* get lock time

* Staking (#31)

* determine lock time

* add mining nullifier

* StartTimeValue

* StartTimeTarget

* set witness

* start time circuit

* fmt

* clippy wip

* single claim proof

* single claim proof

* hash inner

* hash inner circuit

* path change

* fix mock

* rename

* deposit_time_proof

* single claim proof

* wip

* U64 to u64

* save claim circuit

* get lock time

* add chain test

* fixed test

* add eligible check

* faster mining

* Proof of innocence (#32)

* add address list

* innocence inner value

* set witness

* impl cyclic chain circuit

* add test to innocence inner target

* innocence circuit test

* innocence wrap circuit

* innocence processor

* add innocence processor test

* add innocence processor not use allow list

* Sign message function (#33)

* Add sign_message function

* Remove unnecessary copy

* Rename functions

* Fix weight

* add serialize

* impl fromstr

* add default

* fix aggregation proof generation condition (#34)

---------

+++ src/common/trees/account_tree.rs
@@ -47,7 +47,7 @@ impl AccountTree {
         tree
     }
 
-    pub fn prove_inclusion(&self, account_id: usize) -> AccountMerkleProof {
+    pub fn prove_inclusion(&self, account_id: u64) -> AccountMerkleProof {
         let leaf = self.get_leaf(account_id);
         let merkle_proof = self.prove(account_id);
         AccountMerkleProof { merkle_proof, leaf }
@@ -65,7 +65,7 @@ pub struct AccountMerkleProof {
 impl AccountMerkleProof {
     /// id is already registered. Account id range check is assumed already
     /// done.
-    pub fn verify(&self, root: PoseidonHashOut, account_id: usize, pubkey: U256) -> bool {
+    pub fn verify(&self, root: PoseidonHashOut, account_id: u64, pubkey: U256) -> bool {
         let mut result = true;
         let is_not_pubkey_zero = pubkey != U256::default();
         result = result && is_not_pubkey_zero;
